requirements.txt
streamlit==1.48.1
tensorflow==2.20.0
numpy==2.1.3
opencv-python==4.12.0.88
scikit-learn==1.7.1
Pillow==10.4.0
pandas==2.2.3
tqdm==4.67.1
